+++ src/app/api-server/src/http_server.rs
@@ -12,18 +12,17 @@ use std::path::PathBuf;
 use std::sync::Arc;
 
 use database_common_macros::transactional_handler;
-use dill::CatalogBuilder;
 use http_common::ApiError;
 use indoc::indoc;
 use internal_error::{InternalError, ResultIntoInternal};
 use kamu::domain::TenancyConfig;
 use kamu_adapter_http::DatasetAuthorizationLayer;
 use observability::axum::unknown_fallback_handler;
+use tokio::net::TcpListener;
 use tokio::sync::Notify;
 use utoipa_axum::router::OpenApiRouter;
 use utoipa_axum::routes;
 
-use crate::config::UrlConfig;
 use crate::ui_configuration::UIConfiguration;
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -34,7 +33,7 @@ pub async fn build_server(
     catalog: dill::Catalog,
     tenancy_config: TenancyConfig,
     ui_config: UIConfiguration,
-    url_config: UrlConfig,
+    maybe_e2e_http_server_listener: Option<TcpListener>,
     e2e_output_data_path: Option<&PathBuf>,
 ) -> Result<
     (
@@ -44,29 +43,19 @@ pub async fn build_server(
             axum::Router,
         >,
         SocketAddr,
-        dill::Catalog,
         Option<Arc<Notify>>,
     ),
     InternalError,
 > {
     let gql_schema = kamu_adapter_graphql::schema();
 
-    let addr = SocketAddr::from((address, http_port.unwrap_or(0)));
-    let listener = tokio::net::TcpListener::bind(addr).await.int_err()?;
+    let listener = if let Some(listener) = maybe_e2e_http_server_listener {
+        listener
+    } else {
+        let addr = SocketAddr::from((address, http_port.unwrap_or(0)));
+        TcpListener::bind(addr).await.int_err()?
+    };
     let local_addr = listener.local_addr().unwrap();
-
-    let base_url_rest =
-        url::Url::parse(&format!("http://{local_addr}")).expect("URL failed to parse");
-
-    let api_server_catalog = CatalogBuilder::new_chained(&catalog)
-        .add_value(kamu::domain::ServerUrlConfig::new(
-            kamu::domain::Protocols {
-                base_url_rest: base_url_rest.clone(),
-                base_url_platform: url_config.base_url_platform,
-                base_url_flightsql: url_config.base_url_flightsql,
-            },
-        ))
-        .build();
 
     let (mut router, api) = OpenApiRouter::with_openapi(
         kamu_adapter_http::openapi::spec_builder(
@@ -157,7 +146,7 @@ pub async fn build_server(
     .merge(kamu_adapter_http::openapi::router().into())
     .fallback(unknown_fallback_handler)
     .layer(axum::extract::Extension(gql_schema))
-    .layer(axum::extract::Extension(api_server_catalog.clone()))
+    .layer(axum::extract::Extension(catalog.clone()))
     .layer(axum::extract::Extension(ui_config))
     .split_for_parts();
 
@@ -177,12 +166,7 @@ pub async fn build_server(
     let router = router.layer(axum::extract::Extension(std::sync::Arc::new(api)));
 
     let server = axum::serve(listener, router.into_make_service());
-    Ok((
-        server,
-        local_addr,
-        api_server_catalog,
-        maybe_shutdown_notify,
-    ))
+    Ok((server, local_addr, maybe_shutdown_notify))
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////