+++ CHANGELOG.md
@@ -13,11 +13,19 @@ Recommendation: for ease of reading, use the following order:
 -->
 
 ## [Unreleased]
+### Fixed
+- Flow progress notifier is now more resilient to deleted datasets
+- Use actual `base_url` in catalog configuration instead default one
+
+## [0.56.1] - 2025-03-14
+### Fixed
+- Value of flow agent throttling now presented in seconds
+
+## [0.56.0] - 2025-03-13
 ### Added
 - Prometheus metrics: S3 ([kamu CLI `0.226.5`](https://github.com/kamu-data/kamu-cli/releases/tag/v0.226.5))
 - New `FlowSystemConfig` structure in `CLIConfig` which allows to configure `flow_agent` and `task_agent` services with next options `awaiting_step_secs` and `mandatory_throttling_period_secs`
-### Fixed
-- Use actual `base_url` in catalog configuration instead default one
+- GQL: New natural language search API - to use this feature you'll need to configure the OpenAI API key and a [Qdrant](https://qdrant.tech/) vector database connection
 
 ## [0.55.1] - 2025-03-07
 ### Fixed